git-integration/requirements.txt
# Git Integration System Dependencies
# These are separate from the main system to avoid conflicts

requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
